video: Initialize OpenGL
Without this change gpSP will crash with the following error:
/usr/src/debug/wayland/1.23.1/src/wayland-client.c:2460: wl_proxy_set_queue: Assertion `proxy->display == queue->display' failed

This seems to happen after the libSDL 2.30.9 update.

https://github.com/libsdl-org/SDL/commit/0f695adb0392446f36e872e72a828467d5f299f0
removes the initialization of OpenGL if the OpenGL window flag isn't provided.

Simply adding the SDL_WINDOW_OPENGL flag isn't enought to get it to work
again, as it seems that the GL library isn't actually being loaded for
some reason. Instead load the GL library manually.

Signed-off-by: Darrel Griët <dgriet@gmail.com>